insert Shared Preference

+++ app/src/main/java/com/j2d2/insulin/Insulin.kt
@@ -11,7 +11,7 @@ data class Insulin (
     @ColumnInfo(name = "time") val time: String,
     @ColumnInfo(name = "insulin_type") val type: Int?,              // 0:휴물린엔, 1:캐닌슐린
     @ColumnInfo(name = "undiluted_vol") val undiluted: Float?,      // 원액량
-    @ColumnInfo(name = "total_capacity") val totalCapacity: Float?, // 총 주사량(희석일 경우 총량)
+    @ColumnInfo(name = "total_capacity") val totalCapacity: Int?, // 총 주사량(희석일 경우 총량)
     @ColumnInfo(name = "dilution") val dilution: Int?,              // 희석:1, 아님:0
     @ColumnInfo(name = "remark") val remark: String?                // 메모
 )
